Only return fields you're allowed to read

+++ api/src/controllers/fields.ts
@@ -3,7 +3,7 @@ import asyncHandler from 'express-async-handler';
 import FieldsService from '../services/fields';
 import validateCollection from '../middleware/collection-exists';
 import { schemaInspector } from '../database';
-import { FieldNotFoundException, InvalidPayloadException } from '../exceptions';
+import { FieldNotFoundException, InvalidPayloadException, ForbiddenException } from '../exceptions';
 import Joi from 'joi';
 import { Field } from '../types/field';
 import useCollection from '../middleware/use-collection';
@@ -48,7 +48,7 @@ router.get(
 		const service = new FieldsService({ accountability: req.accountability });
 
 		const exists = await schemaInspector.hasColumn(req.collection, req.params.field);
-		if (exists === false) throw new FieldNotFoundException(req.collection, req.params.field);
+		if (exists === false) throw new ForbiddenException();
 
 		const field = await service.readOne(req.params.collection, req.params.field);
 		return res.json({ data: field || null });
@@ -122,8 +122,10 @@ router.patch(
 	useCollection('directus_fields'),
 	// @todo: validate field
 	asyncHandler(async (req, res) => {
-		const service = new FieldsService({ accountability: req.accountability });
+		const exists = await schemaInspector.hasColumn(req.collection, req.params.field);
+		if (exists === false) throw new ForbiddenException();
 
+		const service = new FieldsService({ accountability: req.accountability });
 		const fieldData: Partial<Field> & { field: string; type: typeof types[number] } = req.body;
 
 		if (!fieldData.field) fieldData.field = req.params.field;
@@ -141,6 +143,9 @@ router.delete(
 	validateCollection,
 	useCollection('directus_fields'),
 	asyncHandler(async (req, res) => {
+		const exists = await schemaInspector.hasColumn(req.collection, req.params.field);
+		if (exists === false) throw new ForbiddenException();
+
 		const service = new FieldsService({ accountability: req.accountability });
 
 		await service.deleteField(req.params.collection, req.params.field, req.accountability);
